feat: Add last billed amount to maintenance tasks
Introduces a new field `lastBilledAmount` to the `MaintenanceTask` type to track the last amount billed for a specific task. This change includes updates to the `AddTaskScreen` to allow users to input this value and `TaskDetailsScreen` to display it, formatted as currency. The `lastCompleted` date input is also made a required field in `AddTaskScreen`.